requirements.txt
acres==0.5.0
annotated-types==0.7.0
anyio==4.11.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==25.1.0
babel==2.17.0
beautifulsoup4==4.13.3
bleach==6.2.0
certifi==2025.10.5
cffi==1.17.1
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.3.0
coloredlogs==15.0.1
comm==0.2.2
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dotenv==0.9.9
etelemetry==0.3.1
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.20.0
flatbuffers==25.2.10
fonttools==4.55.8
fqdn==1.5.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
humanfriendly==10.0
idna==3.11
ipykernel==6.29.5
ipython==8.32.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.11.1
joblib==1.4.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
langchain==0.3.27
langchain-core==0.3.75
langchain-text-splitters==0.3.9
langsmith==0.4.18
lightgbm==4.6.0
looseversion==1.3.0
lxml==6.0.2
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mistune==3.1.2
mpmath==1.3.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
notebook==7.3.2
notebook_shim==0.2.4
numpy==2.2.2
onnxruntime==1.22.1
openai==2.6.1
opencv-python==4.11.0.86
orjson==3.11.3
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathlib==1.0.1
patsy==1.0.1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prometheus_client==0.21.1
prompt_toolkit==3.0.50
protobuf==6.32.0
prov==2.1.1
psutil==7.1.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.30
pycparser==2.22
pydantic==2.12.3
pydantic_core==2.41.4
pydot==4.0.1
Pygments==2.19.1
PyMuPDF==1.26.5
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.2.1
python-pptx==1.0.2
pytz==2025.1
pyxnat==1.6.3
PyYAML==6.0.2
pyzmq==26.2.1
rdflib==7.2.1
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.23.1
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==80.9.0
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.43
stack-data==0.6.3
statsmodels==0.14.4
sympy==1.14.0
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.4.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
traits==7.0.2
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.1
uri-template==1.3.0
urllib3==2.3.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.45.1
widgetsnbextension==4.0.13
xlsxwriter==3.2.9
zstandard==0.24.0
